+++ packages/nostrify/CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - NPool: added a `.close()` method and support `await using` syntax.
 
+### Fixed
+
+- NPool: fixed routing of `.req()`.
+
 ## 0.32.0 - 2024-09-20
 
 ### Added